refactor(studio): split remote terrarium startup
Move worker-owned recipe deployment and compensation into a focused lifecycle module, keeping the Studio facade below the hard file-size limit while preserving its private delegate.
